# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для книжной платформы Oqyrman",
        "title": "OqyrmanAPI",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/library-books": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library-books"
                ],
                "summary": "Добавить книгу в библиотеку",
                "parameters": [
                    {
                        "description": "Данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/library_book.createLibraryBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/library_book.libraryBookResponse"
                        }
                    }
                }
            }
        },
        "/admin/library-books/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library-books"
                ],
                "summary": "Обновить копии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/library_book.updateLibraryBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/library_book.libraryBookResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "library-books"
                ],
                "summary": "Удалить запись",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/library-books/book/{book_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library-books"
                ],
                "summary": "Библиотеки с книгой",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID книги",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/library-books/library/{library_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library-books"
                ],
                "summary": "Книги в библиотеке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID библиотеки",
                        "name": "library_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/library-books/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library-books"
                ],
                "summary": "Получить запись",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/library_book.libraryBookViewResponse"
                        }
                    }
                }
            }
        },
...
        "/staff/books/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library-books"
                ],
                "summary": "Поиск книг в библиотеке (для staff)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поиск по названию/автору",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по жанру (UUID)",
                        "name": "genre_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Только доступные",
                        "name": "available",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/library_book.libraryBookSearchResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.AuthorRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "common.BookRef": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/common.AuthorRef"
                },
                "avg_rating": {
                    "type": "number"
                },
                "cover_url": {
                    "type": "string"
                },
                "genre": {
                    "$ref": "#/definitions/common.GenreRef"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "common.GenreRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "common.LibraryRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "library_book.createLibraryBookRequest": {
            "type": "object",
            "required": [
                "available_copies",
                "book_id",
                "library_id",
                "total_copies"
            ],
            "properties": {
                "available_copies": {
                    "type": "integer"
                },
                "book_id": {
                    "type": "string"
                },
                "library_id": {
                    "type": "string"
                },
                "total_copies": {
                    "type": "integer"
                }
            }
        },
        "library_book.libraryBookResponse": {
            "type": "object",
            "properties": {
                "available_copies": {
                    "type": "integer"
                },
                "book_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "library_id": {
                    "type": "string"
                },
                "total_copies": {
                    "type": "integer"
                }
            }
        },
        "library_book.libraryBookSearchItem": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "available_copies": {
                    "type": "integer"
                },
                "book_id": {
                    "type": "string"
                },
                "cover_url": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "library_book_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_copies": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "library_book.libraryBookSearchResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/library_book.libraryBookSearchItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "library_book.libraryBookViewResponse": {
            "type": "object",
            "properties": {
                "available_copies": {
                    "type": "integer"
                },
                "book": {
                    "$ref": "#/definitions/common.BookRef"
                },
                "id": {
                    "type": "string"
                },
                "library": {
                    "$ref": "#/definitions/common.LibraryRef"
                },
                "total_copies": {
                    "type": "integer"
                }
            }
        },
        "library_book.updateLibraryBookRequest": {
            "type": "object",
            "properties": {
                "available_copies": {
                    "type": "integer"
                },
                "total_copies": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}